requirements.txt
python-telegram-bot==13.15
sentence-transformers
faiss-cpu
requests
